add new usage example center-point
