--- public/usage-examples/geometry/circle_at/circle_at-1-simple.cpp
@@ -1,26 +1,0 @@
-#include "splashkit.h"
-
-int main()
-{
-    open_window("Circle At", 800, 600);
-    clear_screen();
-
-    //Set position for the circle
-    double x_position = rand() % 400 + 200;
-    //Give random  x_position value bewteen 200 - 400
-    double y_position = 300;
-
-    //Create a circle at the position (x_position, y_position)
-    circle A = circle_at(x_position, y_position, 200);
-
-    //Draw the Circle
-    draw_circle(COLOR_RED, x_position, y_position, 200);
-    
-    string text = "Circle At: (400,300), Radius: 200";
-    //Print result on window
-    draw_text(text, COLOR_BLACK, 0, 20, 100, 100);
-    
-    refresh_screen();
-    delay(4000);
-    close_all_windows();
-}
