--- public/usage-examples/geometry/circle_at/circle_at-1-simple.txt
@@ -1,3 +1,0 @@
-#### Create circle at specified location
-
-The following code shows examples of using [Circle At](/api/geometry/#circle-at) to show how to create a circle at a user-defined location with a specified radius.
